+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.1] - 2026-01-22
 
 ### Changed
-- Provisioner uses Terraform and Ansible
+- Migrated range and NGFW provisioning from Pulumi to Terraform
 
 ## [1.0.0] - 2026-01-21